@@ -2,7 +2,7 @@ name: beeper_common
 description: The Beeper landing page and console
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.5 <3.0.0'
 
 dependencies:
   clock: ^1.0.1
