@@ -10,13 +10,13 @@ export 'package:beeper/gen/modules.g.dart';
 class Metadata {
   final String name;
   final bool lazyLoad;
-  final Module Function(dynamic data) factory;
+  final Module Function(dynamic data)? factory;
   final bool loadable;
   final bool configurable;
 
   const Metadata({
-    @required this.name,
-    @required this.lazyLoad,
+    required this.name,
+    required this.lazyLoad,
     this.factory,
     this.loadable = false,
     this.configurable = true,
@@ -27,14 +27,14 @@ abstract class ModuleSystem {
   @protected
   var initializing = false;
 
-  ModuleScope scope;
+  late ModuleScope scope;
 }
 
 abstract class Module {
-  ModuleSystem /*!*/ system;
-  ModuleScope /*!*/ scope;
+  late ModuleSystem system;
+  late ModuleScope scope;
 
-  Completer<void> _loaded;
+  Completer<void>? _loaded;
   bool get loaded => _loaded?.isCompleted ?? false;
 
   String get canonicalName {
@@ -43,7 +43,7 @@ abstract class Module {
           (element) => element.value == this,
         )
         .key;
-    final metadata = moduleMetadata[key.item1] /*!*/;
+    final metadata = moduleMetadata[key.item1]!;
     final scopeName = scope.canonicalName;
     if (key.item2 == null) {
       return '$scopeName/${metadata.name}';
@@ -53,25 +53,23 @@ abstract class Module {
   }
 
   Future<void> _performLoad({
-    @required ModuleSystem system,
-    @required ModuleScope scope,
+    required ModuleSystem system,
+    required ModuleScope scope,
   }) async {
     if (_loaded == null) {
       _loaded = Completer();
       this.system = system;
       this.scope = scope;
       await load();
-      _loaded.complete();
+      _loaded!.complete();
     } else {
-      return _loaded.future;
+      return _loaded!.future;
     }
   }
 
   @mustCallSuper
   @protected
-  Future<void> load() async {
-    assert(scope != null);
-  }
+  Future<void> load() async {}
 
   @mustCallSuper
   @protected
@@ -83,24 +81,25 @@ abstract class Module {
 
 class ModuleScope {
   final ModuleSystem system;
-  final ModuleScope parent;
+  final ModuleScope? parent;
 
   ModuleScope({
-    @required this.system,
-    @required this.parent,
+    required this.system,
+    required this.parent,
   });
 
   final _children = <Object, ModuleScope>{};
-  final _modules = <Tuple2<Type, Object>, Module>{};
-  ModuleScope _inherit;
+  final Map<Tuple2<Type, Object?>, Module> _modules =
+      <Tuple2<Type, Object>, Module>{};
+  ModuleScope? _inherit;
 
   Map<Object, ModuleScope> get children => UnmodifiableMapView(_children);
-  Map<Tuple2<Type, Object>, Module> get modules =>
+  Map<Tuple2<Type, Object?>, Module> get modules =>
       UnmodifiableMapView(_modules);
 
-  String _canonicalName;
+  String? _canonicalName;
 
-  String /*!*/ get canonicalName {
+  String get canonicalName {
     if (_canonicalName == null) {
       if (parent == null) {
         _canonicalName = '';
@@ -108,18 +107,18 @@ class ModuleScope {
         final key = children.entries
             .singleWhere((element) => element.value == this)
             .key;
-        _canonicalName = '${parent.canonicalName}/$key';
+        _canonicalName = '${parent!.canonicalName}/$key';
       }
     }
-    return _canonicalName;
+    return _canonicalName!;
   }
 
-  ModuleScope push(Object id) {
+  ModuleScope? push(Object id) {
     if (_children.containsKey(id)) {
       return _children[id];
-    } else if (_inherit._children.containsKey(id)) {
+    } else if (_inherit!._children.containsKey(id)) {
       final scope = ModuleScope(system: system, parent: this);
-      scope._inherit = _inherit._children[id];
+      scope._inherit = _inherit!._children[id];
       _children[id] = scope;
       return scope;
     } else {
@@ -146,22 +145,22 @@ class ModuleScope {
     });
   }
 
-  Module getWith(Type T, {Object id}) {
+  Module? getWith(Type T, {Object? id}) {
     if (T == Module) {
       throw ArgumentError('get requires type argument');
     }
     final key = Tuple2(T, id);
     if (_modules.containsKey(key)) {
-      final module = _modules[key];
-      if (!module._loaded.isCompleted) {
+      final module = _modules[key]!;
+      if (!module._loaded!.isCompleted) {
         throw StateError(
           'Cannot access module that has not finished initialization`',
         );
       }
       return module;
-    } else if (_inherit != null && _inherit._modules.containsKey(key)) {
-      final module = _inherit._modules[key];
-      assert(module._loaded.isCompleted);
+    } else if (_inherit != null && _inherit!._modules.containsKey(key)) {
+      final module = _inherit!._modules[key]!;
+      assert(module._loaded!.isCompleted);
       module.scope = this;
       _modules[key] = module;
       return module;
@@ -170,9 +169,9 @@ class ModuleScope {
     }
   }
 
-  T get<T extends Module /*!*/ >({Object id}) => getWith(T, id: id) as T;
+  T? get<T extends Module>({Object? id}) => getWith(T, id: id) as T?;
 
-  Future<Module> requireWith(Type T, [Object id]) async {
+  Future<Module> requireWith(Type T, [Object? id]) async {
     if (T == Module) {
       throw ArgumentError('require requires type argument');
     } else if (!system.initializing) {
@@ -187,7 +186,7 @@ class ModuleScope {
       throw StateError('Could not find metadata for module type $T');
     }
     if (metadata.lazyLoad) {
-      module = metadata.factory(null);
+      module = metadata.factory!(null);
     } else {
       throw StateError('Module $T not found');
     }
@@ -199,11 +198,11 @@ class ModuleScope {
     return module;
   }
 
-  Future<T> require<T extends Module>({Object id}) async {
+  Future<T> require<T extends Module>({Object? id}) async {
     return await requireWith(T, id) as T;
   }
 
-  Future<void> injectWith(Type T, Module module, {Object id}) async {
+  Future<void> injectWith(Type T, Module module, {Object? id}) async {
     if (T == Module) {
       throw ArgumentError('inject requires type argument');
     } else if (!system.initializing) {
@@ -220,7 +219,7 @@ class ModuleScope {
     );
   }
 
-  Future<void> inject<T extends Module>(Module module, [Object id]) =>
+  Future<void> inject<T extends Module>(Module module, [Object? id]) =>
       injectWith(T, module, id: id);
 
   void dispose() {
