@@ -2,10 +2,9 @@ import 'package:beeper/discord/discord.dart';
 import 'package:beeper/modules.dart';
 import 'package:beeper/modules/status.dart';
 import 'package:beeper/secrets.dart';
-import 'package:meta/meta.dart';
 
 mixin DiscordLoader on Module {
-  Discord/*!*/ discord;
+  late final Discord discord;
 
   @override
   Future<void> load() async {
@@ -16,22 +15,22 @@ mixin DiscordLoader on Module {
 
 @Metadata(name: 'discord')
 class DiscordModule extends Module with StatusLoader {
-  Discord/*!*/ discord;
-
-  final String/*!*/ token;
-  final String endpoint;
+  final String token;
+  final String? endpoint;
 
   DiscordModule({
-    @required this.token,
+    required this.token,
     this.endpoint,
   });
+
+  late final Discord discord;
 
   @override
   Future<void> load() async {
     await super.load();
     discord = Discord(
       token: decryptSecret('discord-token', token),
-      endpoint: endpoint == null ? null : Uri.parse(endpoint),
+      endpoint: endpoint == null ? null : Uri.parse(endpoint!),
     );
     await discord.start();
 
@@ -41,10 +40,10 @@ class DiscordModule extends Module with StatusLoader {
         'guilds': discord.guilds.length,
         if (discord.user != null)
           'user': {
-            'id': discord.user.id,
-            'name': discord.user.name,
-            'discriminator': discord.user.discriminator,
-            'avatar': discord.user.avatar().toString(),
+            'id': discord.user!.id,
+            'name': discord.user!.name,
+            'discriminator': discord.user!.discriminator,
+            'avatar': discord.user!.avatar().toString(),
           },
       };
     });
